added spring data jpa support

+++ src/main/java/org/serhii/example/query/PlayersQueryResolver.java
@@ -17,7 +17,7 @@ public class PlayersQueryResolver implements GraphQLQueryResolver {
         return playerRepository.findAll();
     }
 
-    public Player player(Integer id) {
+    public Player player(Long id) {
         return playerRepository.findOne(id);
     }
